* changes:
  fix(runtime): fix MISRA C:2012 Rule 10.1 in run.c
  fix(runtime): fix MISRA C:2012 Rule 10.1 in handler.c
  fix(lib/xlat): fix MISRA C:2012 Rule 10.1 in xlat_contexts.c
  fix(runtime): fix MISRA C:2012 Rule 10.1 in exit.c
  fix(lib/rmm_el3_ifc): fix MISRA C:2012 Rule 10.1
  fix(lib/arch): fix MISRA C:2012 Rule 10.1 in atomics.h
  fix(lib/realm): fix MISRA C:2012 Rule 10.1,4 in buffer.c
  fix(lib/realm): fix MISRA C:2012 Rule 10.1 in s2tt.c
  fix(lib/arch): fix MISRA C:2012 Rule 10.1 in simd.c
  fix(lib/arch): fix MISRA C:2012 Rule 10.1 in pmu.c
  fix(lib/arch): fix MISRA C:2012 Rule 10.1 in atomics.h
  fix(rmm): fix MISRA C:2012 Rule 8.3 violations
  fix(lib/realm): fix MISRA violations in headers
  fix(lib/attestation): fix MISRA C:2012 defects in attestation_token.c
  fix(runtime): fix MISRA-C:2012 Rule 8.6 in exit.c
  fix(drivers/pl011): fix MISRA-C:2012 Rule 10.1 in pl011.c

+++ lib/xlat/src/xlat_tables_private.h
@@ -53,7 +53,7 @@ void xlat_tables_print(struct xlat_ctx *ctx);
 /*
  * Returns a block/page table descriptor for the given level and attributes.
  */
-uintptr_t xlat_desc(uint64_t attr, uintptr_t addr_pa, int level);
+uint64_t xlat_desc(uint64_t attr, uintptr_t addr_pa, int level);
 
 /*
  * Return the maximum physical address supported by the hardware.
